Refactor and split components
